+++ CHANGELOG.md
@@ -6,6 +6,16 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Upcoming]
 
+## Added
+
+- Added UV index to weather information display (only supported by World Weather Online provider for now).
+- Added dew point to weather information (calculated from temperature and humidity by Magnus formula).
+- Added `--align-right` CLI option to align labels to the right in text output (as `align_right` config option).
+
+## Changed
+
+- Changed default layout, now precipitation is shown in separate line, humidity and dew point are shown together.
+
 ## Fixed
 
 - Fixed bug with World Weather Online not working in different languages.